+++ HISTORY.rst
@@ -8,7 +8,6 @@ Contributors to this version: Pascal Bourgault (:user:`aulemahal`), Éric Dupuis
 
 New features and enhancements
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-* McArthur Forest Fire Danger Index and related indices (:issue:`1152`, :pull:`1159`)
 * Adjustment methods of `SBCK <https://github.com/yrobink/SBCK>`_ are wrapped into xclim when that package is installed. (:issue:`1109`, :pull:`1115`).
     - Wrapped SBCK tests are also properly run in the tox testing ensemble. (:pull:`1119`).
 * Method ``FAO_PM98`` (based on Penman-Monteith formula) to compute potential evapotranspiration. (:pull:`1122`).
@@ -18,7 +17,7 @@ New features and enhancements
 
 New indicators
 ^^^^^^^^^^^^^^
-* `xclim` now has McArthur Forest Fire Danger Index and related indicators under a new ``xclim.indices.fire`` module. Also available as an Indicator. (:issue:`1152`, :pull:`1159`)
+* `xclim` now has the McArthur Forest Fire Danger Index and related indices under a new ``xclim.indices.fire`` module. These indices are also available as indicators. (:issue:`1152`, :pull:`1159`)
 * Drought-related indicators: SPI (standardized precipitations) and SPEI (standardized water budgets). (:issue:`131`, :pull:`1096`).
 * ``ensembles.create_ensembles`` now accepts a ``realizations`` argument to assign a coordinate to the "realization" axis. It also accepts a dictionary as input so that keys are used as that coordinate. (:pull:`1153`).
 * ``ensembles.ensemble_percentiles``, ``ensembles.ensemble_mean_std_max_min`` and ``ensembles.change_significance`` now support weights (:pull:`1151`).